Update test preparer
Changing mSIM configuration no longer causes reboot, so the test
preparer now will trigger reboot and wait for device to come back up.
(This should not be needed eventually, as the radio config can
change now without rebooting, but this is blocked by b/150486156
for now, so we temporarily will just trigger the reboot to be sure).

Bug: 148897549
Test: atest VtsHalRadioV1_2TargetTest
Change-Id: If1f5d72be09e30338d4736ea5d0a265e1fc2a010