VtsFastbootVerification: allow IPv6 link-local address.
The fastbootd over ethernet was implemented with IPv6 link-local
address. An IPv6 address may include many ':'. It will break the fuzzy
test. And because the IPv6 addresses in normal mode can be randomized,
that address cannot be used to connect to the device in the fastbootd
mode.

This change enables IPv6 address as device serial. And it allows users
to set a mapping from device serial to fastbootd serial through
environment  variables to solve the addresses difference between normal
mode and fastbootd mode. And it will format the serial for TCP fastboot
commands with a "tcp:" prefix, like the host fastboot command.

BUG: 157887327
Test: export FASTBOOT_SERIAL_192_168_1_10_5555= \
        "fe80::230:1bff:feba:8128%en0:5555"
      vts-tradefed run vts -s 192.168.1.10:5555 \
        -m VtsFastbootVerification

Change-Id: Ie7a1443dfa534168d579c955887254f4fea687d9