[b/168100438]Fix VtsFastbootVerification failure for N/W adb devices
For non-usb client devices, ethernet in recovery mode and fastbootd
via network patch was backported from Android 11 to Android 10
https://android-review.googlesource.com/q/topic:%22
fastbootd_over_ethernet%22+(status:open%20OR%20status:merged)

After this backport, when VtsFastbootVerification testcase was run
the self diagonosis framework of vts marks the testcase as failure,
in two instances when it checks adb connection it marks it as failure,
because adb is not yet initialised completely and adb authorisation
is still in progress so adb devices command shows offline,
which is treated as an error by testsuite.Same check is performed
for all other testcases in the module,so the testcases gets failed.

To mitigate this issue, added the polling criteria in testcase,
which waits till the device is properly initialised, in our case,
it waits for adb to enter into unauthorized state and then the
testcase proceeds as usual. It(polling criteria) is needed at two
places one when the testcase module is initialised & other one after
the first testcase is run(VtsFastbootVerification#testFastbootReboot)

Signed-off-by: Sai Mukkundan <saimukkundan.ramamoorthy@technicolor.com>
Change-Id: I335cf0f6e14d149b4204f8477ebed11b6bc40c57